Change the loops to use the list index instead in task 4

+++ 100-clean_web_static.py
@@ -88,7 +88,10 @@ def do_clean(number=0):
     n = int(number)
     if n == 0 or n == 1:
         n = 1
-    for i in range(0, len(file_names) - n):
-        local("rm versions/{}".format(file_names[i]))
+    for i in file_names[n:]:
+        local("rm versions/{}".format(i))
+    for i in dir_server_names[n:]:
+        if i == 'test':
+            continue
         run("rm -rf /data/web_static/releases/{}"
-            .format(dir_server_names[i]))
+            .format(i))
